Image Classifier/requirements.txt
tensorflow>=2.8.0
pillow
numpy
tkinterdnd2
